requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic-settings==2.10.1
pyyaml==6.0.2
uvicorn==0.35.0
